+++ CHANGELOG.md
@@ -6,12 +6,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [unreleased]
+## [0.2.3]
 
 ### Added
 
 - Basic API for EDAC functionality
 - PWM implementation and example
+
+### Changed
+
+- Improved Timer API. It is now possible to simply use `new` on `CountDownTimer`
 
 ## [0.2.2]